Hide `Mark Read` button if list is for already read
In the case of already read notification does not make sense to have the
button for mark and done/read. In GH terms they use done/read as
similar meaning, but they have also `archive` for saying to remove the
notification from the list of them. But that API at the moment is in
beta, let's see if I can rev eng that.

Signed-off-by: Federico Guerinoni <guerinoni.federico@gmail.com>